further refinements and adding sky

+++ src/JamTemplate/Lerp.hpp
@@ -20,7 +20,7 @@ namespace JamTemplate
 		static T cosine(T const& a, T const& b, T const& t)
 		{
 			assert(t >= 0 && t <= 1);
-			float tRemapCosine = (1.0 - cos(t * 3.1415926f)) * 0.5f;
+			float tRemapCosine = (1.0f - static_cast<float>(cos(t * 3.1415926f))) * 0.5f;
 			return linear(a, b, tRemapCosine);
 		}
 	};
